testing

+++ java/testing/contracts/src/main/java/com/ggar/hibiki/test/contracts/library/ArtistDeletedCleanupEventHandlerContractTest.java
@@ -1,15 +1,14 @@
 package com.ggar.hibiki.test.contracts.library;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.ggar.hibiki.core.catalog.event.ArtistDeletedEvent;
 import com.ggar.hibiki.test.support.ScenarioResult;
+import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 public abstract class ArtistDeletedCleanupEventHandlerContractTest {
 
@@ -23,7 +22,7 @@ public abstract class ArtistDeletedCleanupEventHandlerContractTest {
                 .artistId(artistId)
                 .name("Old Artist")
                 .build();
-        
+
         ScenarioResult<Void> result;
 
         @BeforeEach
@@ -34,8 +33,8 @@ public abstract class ArtistDeletedCleanupEventHandlerContractTest {
         @Test
         @DisplayName("then library items for this artist should be removed")
         final void thenRemoved() {
-            assertThat(result.getState().get("itemsFoundBefore")).isGreaterThan(0);
-            assertThat(result.getState().get("itemsFoundAfter")).isEqualTo(0);
+            assertThat((Integer) result.getState().get("itemsFoundBefore")).isGreaterThan(0);
+            assertThat((Integer) result.getState().get("itemsFoundAfter")).isEqualTo(0);
         }
     }
 }
